use cached downloads

+++ HEFS_Reader/Implementations/HEFS_Downloader.cs
@@ -10,26 +10,45 @@ namespace HEFS_Reader.Implementations
 	{
 		private const string _rootUrl = "https://www.cnrfc.noaa.gov/csv/";
 		public string Response { get; set; }
-		public bool FetchData(HEFSRequestArgs args)
+		public bool FetchData(HEFSRequestArgs args, string cacheDirectory = "")
 		{
             //https://www.cnrfc.noaa.gov/csv/2019092312_RussianNapa_hefs_csv_hourly.zip
             string webrequest = _rootUrl;
-            webrequest += args.date + "_";
-            webrequest += args.location.ToString();
-			webrequest += "_hefs_csv_hourly.zip";
 
-            string zipFileName = Path.GetTempFileName();
-            string csvFileName = Path.GetTempFileName();
+            if (cacheDirectory == "")
+                cacheDirectory = Path.GetTempPath();
+             string fileName  = args.date + "_";
+            fileName += args.location.ToString();
+            fileName += "_hefs_csv_hourly";
+            webrequest += fileName+".zip";
 
+            
+            string zipFileName = Path.Combine(cacheDirectory, fileName+".zip");
+            string csvFileName = Path.Combine(cacheDirectory, fileName + ".csv");
+            if (File.Exists(csvFileName))
+            {
+                Console.WriteLine("Found "+ csvFileName+" in cache.  Reading...");
+                Response = File.ReadAllText(csvFileName);
+                return true;
+            }
+           
+                Console.WriteLine("GET "+webrequest);
             File.Delete(zipFileName);
             File.Delete(csvFileName);
-            GetFile(webrequest, zipFileName);
-
+            try
+            {
+                GetFile(webrequest, zipFileName);
+            }
+            catch( Exception exception)
+            {
+                Console.WriteLine("download failed");
+                File.Delete(zipFileName);
+                File.Delete(csvFileName);
+                return false;
+            }
             Reclamation.Core.ZipFileUtility.UnzipFile(zipFileName, csvFileName);
             Response = File.ReadAllText(csvFileName);
-            File.Delete(zipFileName);
-            File.Delete(csvFileName);
-
+            Console.WriteLine("sucessfully downloaded to "+csvFileName);
 
 
             return true;
